+++ twitchio/parse.py
@@ -69,8 +69,19 @@ def parser(data: str, nick: str):
         message = " ".join(groups[4:]).lstrip(":")
         user = re.search(USER_SUB, groups[1]).group("user")
 
+    elif groups[2] == "USERNOTICE":
+        action = groups[2]
+        channel = groups[3].lstrip("#")
+        message = " ".join(groups[4:]).lstrip(":")
+
     elif action in ACTIONS:
         channel = groups[-1].lstrip("#")
+
+    elif groups[3] in {"PRIVMSG", "PRIVMSG(ECHO)"}:
+        action = groups[3]
+        channel = groups[4].lstrip("#")
+        message = " ".join(groups[5:]).lstrip(":")
+        user = re.search(USER_SUB, groups[2]).group("user")
 
     if action in ACTIONS2:
         prebadge = groups[0].split(";")