+++ docs/changelog.rst
@@ -9,6 +9,10 @@ New logo!
     - Re-add ``raw_usernotice`` from V1.x
     - Fix echo messages for replies
     - Fix a bug where the wrong user would be whispered
+    - Add support for team and channelteams API routes
+        - :class:`Team`, :class:`ChannelTeams`
+        - :func:`Client.fetch_teams`
+        - :func:`PartialUser.fetch_channel_teams`
 
 2.1.1
 ======